feat: auto fold players after 30 seconds inactive in bet time

+++ packages/events/src/modules/poker/timerManager.ts
@@ -1,8 +1,12 @@
+/* eslint-disable @typescript-eslint/no-non-null-assertion */
 import { bot } from '../..';
+import commandRepository from '../../database/repositories/commandRepository';
 import pokerRepository from '../../database/repositories/pokerRepository';
 import { getOrchestratorClient } from '../../structures/ipcConnections';
-import { cleanupGame } from './handleGameAction';
-import { DeleteMatchTimer, PokerTimer, TimerActionType } from './types';
+import { DatabaseCommandSchema } from '../../types/database';
+import { closeTable, updateGameState } from './handleGameAction';
+import PokerFollowupInteractionContext from './PokerFollowupInteractionContext';
+import { DeleteMatchTimer, PokerTimer, TimeoutFoldTimer, TimerActionType } from './types';
 
 const timers = new Map<string, NodeJS.Timeout>();
 
@@ -12,15 +16,45 @@ const executeDeleteMatch = async (timer: DeleteMatchTimer) => {
 
   if (match.inMatch) return;
 
-  cleanupGame(match);
+  const pokerCommandId = (await commandRepository.getCommandInfo('poker')) as DatabaseCommandSchema;
+
+  const ctx = new PokerFollowupInteractionContext(match.interactionToken, pokerCommandId.discordId);
+
+  closeTable(ctx, match);
+};
+
+const executeFoldTimeout = async (timer: TimeoutFoldTimer) => {
+  const match = await pokerRepository.getMatchState(timer.matchId);
+  if (!match) return;
+
+  if (!match.inMatch) return;
+
+  const player = match.players.find((a) => a.seatId === match.seatToPlay)!;
+
+  player.pot = 0;
+  player.folded = true;
+
+  match.lastAction = {
+    action: 'FOLD',
+    playerSeat: player.seatId,
+    pot: match.lastAction.pot,
+  };
+
+  const pokerCommandId = (await commandRepository.getCommandInfo('poker')) as DatabaseCommandSchema;
+
+  const ctx = new PokerFollowupInteractionContext(match.interactionToken, pokerCommandId.discordId);
+
+  return updateGameState(ctx, match);
 };
 
 const executeTimer = async (timerId: string, timer: PokerTimer): Promise<void> => {
-  clearTimer(timerId);
+  clearPokerTimer(timerId);
 
   switch (timer.type) {
     case TimerActionType.DELETE_GAME:
       return executeDeleteMatch(timer);
+    case TimerActionType.TIMOEUT_FOLD:
+      return executeFoldTimeout(timer);
   }
 };
 
@@ -36,7 +70,12 @@ const setupTimers = async (): Promise<void> => {
 
 const startPokerTimeout = (timerId: string, timerMetadata: PokerTimer): void => {
   if (!bot.isMaster) {
-    getOrchestratorClient().send({ type: 'BE_MERCURY', timerId, timerMetadata });
+    getOrchestratorClient().send({
+      type: 'BE_MERCURY',
+      action: 'SET_TIMER',
+      timerId,
+      timerMetadata,
+    });
     return;
   }
 
@@ -44,12 +83,17 @@ const startPokerTimeout = (timerId: string, timerMetadata: PokerTimer): void => 
 
   const timeout = setTimeout(() => {
     executeTimer(timerId, timerMetadata);
-  }, Date.now() - timerMetadata.executeAt).unref();
+  }, timerMetadata.executeAt - Date.now()).unref();
 
   timers.set(timerId, timeout);
 };
 
-const clearTimer = (timerId: string): void => {
+const clearPokerTimer = (timerId: string): void => {
+  if (!bot.isMaster) {
+    getOrchestratorClient().send({ type: 'BE_MERCURY', action: 'CLEAR_TIMER', timerId });
+    return;
+  }
+
   pokerRepository.deleteTimer(timerId);
 
   const timer = timers.get(timerId);
@@ -59,4 +103,4 @@ const clearTimer = (timerId: string): void => {
   timers.delete(timerId);
 };
 
-export { clearTimer, startPokerTimeout, setupTimers };
+export { clearPokerTimer, startPokerTimeout, setupTimers };
